+++ R/artif_utils.R
@@ -38,7 +38,10 @@ add_intervals_from_artifacts <- function(old_events, artifact_found, sample_rang
             imap_dtr( ~{
                 if(all(.x[!is.na(.x)]==FALSE)){
                   out <- new_events_tbl()  
-                    out[,.id:=NULL][]  ## I need to remove .id because it gets added by map  
+                    out[,.id:=NULL][  ## I need to remove .id because it gets added by map 
+                       ,.initial :=  sample_int(integer(0),
+                                                  sampling_rate =
+                                                      sampling_rate(old_events))][]
                 } else {
                     ## left and right values of the window of bad values (respecting the min max samples)
                     left <-  .eeg$.sample_id[.x] + sample_range[[1]]
@@ -53,16 +56,20 @@ add_intervals_from_artifacts <- function(old_events, artifact_found, sample_rang
                        by=.(group=cumsum(c(1, tail(start, -1) > head(stop, -1))))] 
                     data.table::data.table(type = "artifact",
                                            description=type,
-                                           .sample_0 = intervals$start,
-                                           .size = intervals$stop + 1L - intervals$start,
+                                           .initial = sample_int(intervals$start,
+                                                                  sampling_rate =
+                                                                      sampling_rate(old_events)),
+                                           .final = sample_int(intervals$stop,
+                                                               sampling_rate =
+                                                                   sampling_rate(old_events)),
                                            .channel = .y)
                 }
             }),.id = TRUE
             )
     events_found[,.id:= as.integer(.id)]
+
     new_events <- rbind(events_found, old_events, fill = TRUE)
-    
-    data.table::setorder(new_events,.id, .sample_0, .channel)
+    data.table::setorder(new_events,.id, .initial, .channel)
     new_events
 }