+++ R/to_eeg_lst.R
@@ -101,14 +101,13 @@ as_eeg_lst.mne.io.base.BaseRaw <- function(.data){
                                         #create events object
     ann <- .data$annotations$`__dict__`
     if(length(ann$onset)==0){
-        new_events <-new_events_tbl()
+        new_events <-new_events_tbl(.initial =
+                                        sample_int(integer(0), sampling_rate = .data$info$sfreq))
     } else{
     new_events <- new_events_tbl(.id = 1L,
-                             .sample_0 = ann$onset %>%
-                                 as_sample_int(sampling_rate = .data$info$sfreq, unit="s") %>% as.integer,
-                             .size = ann$duration %>%
-                                 {as_sample_int(. ,sampling_rate = .data$info$sfreq,unit="s")-1L} %>%
-                                 as.integer,
+                             .initial = ann$onset %>%
+                                 as_sample_int(sampling_rate = .data$info$sfreq, unit="s"),
+                             .final = as_sample_int(ann$onset + ann$duration,sampling_rate = .data$info$sfreq,unit="s") - 1L,
                              .channel = NA_character_,
                              descriptions_dt = data.table::data.table(description = ann$description))
     }