+++ R/zzz.R
@@ -6,7 +6,7 @@
 		  "..cols_signal_temp",".GRP",".I",".N",".SD",".lower",".new_id",".sid",".upper",".x",".y",
 		  "BinaryFormat","DataFile","DataFormat","DataOrientation","L","MarkerFile",
 		  "Mk_number=Type","SamplingInterval","V1","V2","amplitude",
-		  "channel","","i..sample_0",
+		  "channel","","i..initial",
 		  "i..size","lowerb","mk","n","offset","","recording","resolution",
 		  "scale_fill_gradientn","time","type","value","x..lower","x..sample_id"
 		  )